Add a `--kube-version` option to pass to `helm template`
Currently, templates are only generated with the current Kubernetes
version, but it is possible to validate against any schema. This
means that it's not actually possible to generate what you intend
to validate. This commit adds `--kube-version` option, which gets
passed down to `helm template`.

This option is deliberately separate from `--kubernetes-version`
because a) they are passed to two different commands and b) there
may be reasons why a user might want to validate against a schema
that doesn't match what the template was generated for, such as to
validate cross-version compatibility.